Removed custom modules and extend Custom_item with multiple selection dropdown
