+++ composer.json
@@ -16,8 +16,7 @@
     "psr-4": {
       "PublishPress\\Checklists\\Core\\": "core/",
       "PublishPress\\Checklists\\Permalinks\\": "modules/permalinks/lib",
-      "PublishPress\\Checklists\\Yoastseo\\": "modules/yoastseo/lib",
-      "PublishPress\\Checklists\\Customs\\": "modules/customs/lib"
+      "PublishPress\\Checklists\\Yoastseo\\": "modules/yoastseo/lib"
     }
   },
   "require": {
