+++ core/Requirement/Custom_item.php
@@ -13,7 +13,7 @@ use PPCH_Checklists;
 
 defined('ABSPATH') or die('No direct script access allowed.');
 
-class Custom_item extends Base_simple implements Interface_required
+class Custom_item extends Base_multiple implements Interface_required
 {
     const VALUE_YES = 'yes';
 
@@ -48,8 +48,11 @@ class Custom_item extends Base_simple implements Interface_required
      */
     public function init_language()
     {
-        $this->lang['label']          = __('Custom', 'publishpress-checklists');
-        $this->lang['label_settings'] = __('Custom', 'publishpress-checklists');
+        $this->lang['label']               = __('Custom', 'publishpress-checklists');
+        $this->lang['label_settings']      = __('Custom', 'publishpress-checklists');
+        $this->lang['label_option_title']  = __('Permitted roles', 'publishpress-checklists');
+        $this->lang['label_not_permitted'] = __('Sorry, this post needs the approval of a user in role %s', 'publishpress-checklists');
+        $this->lang['label_approved']      = __('Approved', 'publishpress-checklists');
     }
 
     /**
@@ -141,15 +144,22 @@ class Custom_item extends Base_simple implements Interface_required
         // Enabled
         $enabled = $this->is_enabled();
 
+        //set custom to false if user role is not permitted to prevent any validation
+        if ($this->isUserRolePermitted()) {
+            $is_custom = true;
+        } else {
+            $is_custom = false;
+        }
+
         // Register in the requirements list
         if ($enabled) {
             $requirements[$this->name] = [
                 'status'    => $this->get_current_status($post, $enabled),
-                'label'     => $this->get_title(),
+                'label'     => $this->get_requirement_drop_down_label($post->ID),
                 'value'     => $enabled,
                 'rule'      => $rule,
                 'id'        => $this->name,
-                'is_custom' => true,
+                'is_custom' => $is_custom,
                 'type'      => $this->type,
             ];
         }
@@ -179,12 +189,22 @@ class Custom_item extends Base_simple implements Interface_required
      */
     public function filter_settings_validate($new_options)
     {
+        // Make sure to remove the options that were cleaned up
+        foreach ($new_options as $key => $value) {
+            if (preg_match('/_multiple$/', $key)) {
+                if (!isset($_POST['publishpress_checklists_checklists_options'][$key])) {
+                    unset($new_options[$key]);
+                }
+            }
+        }
+
         if (isset($new_options[$this->name . '_title'][$this->post_type])
             && empty($new_options[$this->name . '_title'][$this->post_type])) {
             // Look for empty title
             $index = array_search($this->name, $new_options['custom_items']);
             if (false !== $index) {
                 unset(
+                    $new_options[$this->name . '_multiple'][$this->post_type],
                     $new_options[$this->name . '_title'][$this->post_type],
                     $new_options[$this->name . '_rule'][$this->post_type],
                     $new_options['custom_items'][$index]
@@ -196,6 +216,9 @@ class Custom_item extends Base_simple implements Interface_required
         if (isset($new_options['custom_items_remove'])
             && !empty($new_options['custom_items_remove'])) {
             foreach ($new_options['custom_items_remove'] as $id) {
+                $var_name = $id . '_multiple';
+                unset($new_options[$var_name]);
+
                 $var_name = $id . '_title';
                 unset($new_options[$var_name]);
 
@@ -214,5 +237,96 @@ class Custom_item extends Base_simple implements Interface_required
         unset($new_options['custom_items_remove']);
 
         return $new_options;
+    }
+
+    /**
+     * Check if user role is permitted to validate this task
+     */
+    private function isUserRolePermitted()
+    {
+
+        // Option name
+        $option_name_multiple = $this->name . '_multiple';
+
+        //Saved value
+        $option_value = isset($this->module->options->{$option_name_multiple}[$this->post_type]) ? $this->module->options->{$option_name_multiple}[$this->post_type] : array();
+
+        if (!isset($this->module->options->{$option_name_multiple}[$this->post_type])) {
+            return true;
+        }
+
+        if (array_intersect($option_value, wp_get_current_user()->roles)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Gets the requirement drop down label.
+     * @param integer $post_id
+     *
+     * @return string
+     */
+    public function get_requirement_drop_down_label($post_id)
+    {
+        //Option name
+        $option_name_multiple = $this->name . '_multiple';
+
+        //Saved value
+        $option_value = isset($this->module->options->{$option_name_multiple}[$this->post_type]) ? $this->module->options->{$option_name_multiple}[$this->post_type] : array();
+
+        //Permitted role list
+        $permitted_role = $this->arrayToSentence($option_value);
+
+        if ($this->isUserRolePermitted()) {
+            $label = $this->get_title();
+        } elseif (self::VALUE_YES === get_post_meta($post_id, PPCH_Checklists::POST_META_PREFIX . $this->name, true)) {
+            $label = $this->lang['label_approved'];
+        } else {
+            $label = sprintf(__($this->lang['label_not_permitted'], 'publishpress-checklists'), $permitted_role);
+        }
+
+        return $label;
+    }
+
+    /**
+     * Form readable sentence from array list
+     * @param array $array
+     *
+     * @return string
+     */
+    private function arrayToSentence($array, $string = '')
+    {
+        if (is_array($array)) {
+            $items = array();
+
+            foreach ($array as $item) {
+                $items[] = $this->get_setting_drop_down_labels()[$item];
+            }
+
+            $last = array_pop($items);
+
+            $string = count($items) ? implode(", ", $items) . " " . __('or', 'publishpress-checklists') . " " . $last : $last;
+        }
+
+        return $string;
+    }
+
+    /**
+     * Gets settings drop down labels.
+     *
+     * @return array.
+     */
+    public function get_setting_drop_down_labels()
+    {
+        $labels = [];
+
+        $userRoles = get_editable_roles();
+
+        foreach ($userRoles as $slug => $role) {
+            $labels[$slug] = $role['name'];
+        }
+        return $labels;
     }
 }
